fix
git-svn-id: http://svn.umitproject.org/svnroot/umit/branch/PacketManipulator@3399 1105ee14-b0fa-0310-85a5-ff3eed429ff7

+++ widgets/Sniff.py
@@ -266,7 +266,7 @@ class SniffPage(gtk.VBox):
         return False
 
     def __on_stop(self, action):
-        if self.context.is_alive():
+        if getattr(self, 'context', None) and self.context.is_alive():
             self.context.destroy()
 
     def __on_save(self, action, saveas_on_fail=True):
